stationary things in one file
